Split tests; use Python's unittest discovery
