+++ run_tests.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+/usr/bin/env python -m unittest discover
